Create the ESRGAN model and necessary files
First commit to create the base files for the ESRGAN. Minor changes to
dataset.py since we are using a different data structure, organised in
low/high resolution subfolders as opposed to a single HR folder from
which the LR images are produced by down-scaling. Change the default LR
and HR shapes to 20x20 and 80x80.

[Ticket: 1]