Add a loss function based on the brightness
The bright_loss is a simple MAE loss where the inputs are the surface
brightness of the fake and the target images. The surface brightness is
calculated summing the pixels of the image. Since the two images have
the same pixel density, the two sum are compared directly without prior
scaling.

[Ticket: X]