Add the model ESRT to improve the inference time
The inference time of the previous model (ESRGAN) was ~0.1 s/image on a
K80 GPU. This is not useful if the dataset is composed by >100k images.
Therefore we introduce here the Efficient Super Resolution Transformer
(ESRT) model, a lightweight implementation of the transformer which
employs the encoder module of the original Transformer coupled with a
SRCNN-like module. This DNN allows an inference time up to ~20x faster.

[Ticket: X]