Add a new alternative model (FSRCNN)
The poor quality of the ESRT compared to ESRGAN is probably due to the
complexity of tuning such a complex deep net, and not due to depth of
the model itself. On the contrary, ESRT may still be too complex for
the task. FSRCNN is a lightweight (~13k parameters) net that produces
high quality SR images (comparable with the output of the ESRGAN, at
least regarding our sample), it less prone to overfitting and could
achive 1000x inference speed compared to the ESRGAN.

[Ticket: X]