Modify the net to best fit the new dataset
The new considered dataset has 5 parameters and it is linearly sampled
in those parameters. It seems that removing all the intermediate layers
in the net provides better results. This is an unexpected behavior, but
can be used to improve the architecture by adding a skip connection
between the first convolution block and the upsampling blocks. This
small addition improved the loss by a factor of 2.

[Ticket: X]